Improve `toBeStable()`
